Don't quote "update_after" in "stale" argument to view.
stale can be true or false which are usually passed in as bools and the
right thing happens, but you can also have stale=update_after which
should come in as a string, but not be quoted.

We could also make a special type for that, but this is easier to use.

cbugg: close bug-606