Tap feed support
Bucket.StartTapFeed starts a feed with a channel you can read events
from.

Edge case: the feed will not detect if a new node joins the cluster,
and won't hook up to its tap feed and relay events. It should however
work correctly if nodes leave a cluster or if the nodes load-balance.