Handle auth on bucket requests
This has me shoving credentials into the requests manually, working
around a bug in go's stdlib net/url that caused it to encode the
credentials correctly.

It also revealed an issue with the server where I couldn't find a bucket
in a list even when authenticated if that bucket required a password.

cbugg: close bug-771