Expose connection pool timeout
Declared ConnPoolTimeout as an exportable variable. This will enable the
programs that use this client to set it to a value that is
required. Programs that use this can also perform some actions upon
getting the timeout. Proposed workaround for the problem specified
in cbugg [bug-774](http://cbugg.hq.couchbase.com/bug/bug-774).

Whenever, HttpClient is used to make requests, timeout can be specified
by setting HttpClient to a custom http.Client as mentioned
in [this StackOverflow question](http://stackoverflow.com/a/16895878/1043536)