cbdatasource: handle stopCh and bad vbucket map
These fixes were found when testing bucket FLUSH scenarios...

The stopCh is closed when the application wants to close the
cbdatasource, and this change adds another place where the stopCh
should be checked.

Also, during a FLUSH, the vbucket map provided by couchbase server
might be temporarily bad (such as an empty vbucket map), so this
change also kicks the cbdatasource to "refresh from the top" when we
detect this situation.

Change-Id: I8868d796974fd04a8193b499203b1f229b5bac2b
Reviewed-on: http://review.couchbase.org/57740
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>