MB-17540 - cbdatasource refreshClusterReasons kick deduplication
During startup of a couchbase server instance, when cbdatasource is
used by an integrated service (like FTS in watson), the cbdatasource
library can be engaged quite early.  So, cbdatasource might see lots
of initial not-my-vbucket errors as ep-engine is still initializing
and warming up its vbuckets.

This is seen in cbdatasource stats...

    "TotUPRStreamReqResFailNotMyVBucket": 4773,

And that leads to lots of "refresh cluster" kicking in cbdatasource.

With this change, there's better deduplication of the kicking...

    "TotKick": 4774,
    "TotKickDeduped": 4717,
    "TotKickOk": 57,

Each deduped kick means one less REST request to ns-server.  Above,
only 57 kicks resulted in a ns-server REST request rather than
thousands.

Change-Id: Ibffac95630b9c04549035ec00bf137dcdb1fb570
Reviewed-on: http://review.couchbase.org/59211
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>