MB-17996 - use GenericMcdAuthHandler auth if available
The GenericMcdAuthHandler interface (usually implemented by cbauth) is
the correct implementation of the same idea as the short-lived
ServerCredProvider interface.

One minor difference is the GenericMcdAuthHandler also selects the
right bucket on the provided client object.

Related, if we auth based on a couchbase.AuthWithSaslHandler approach,
we should not try to select-bucket (as the memcached server responds
with error if we try a select-bucket on a SASL auth'ed connection).

Change-Id: I97e7a42dbcfd0959e9b114cfba2a0281c62c9297
Reviewed-on: http://review.couchbase.org/60364
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>