MB-27415 n1ql churns too many TCP connections
When the n1ql service gets itself into overflow connection territory,
as soon as an overflow connection is done, it gets closed straight away,
however, for hard core Q1, a new one will be required straight away, which
leads to lots connections in TIME_WAIT and depletion of local ports.
If we can't return a connection to the pool straight away, we'll fire a
goroutine which waits 1ms before closing the connection.
If we manage to return to the pool the connection within that timeframe, we'll
keep juggling overflow connections without getting into
Software Induced Oscillations (have you ever heard of PIOs in avionics? ;-)
and avoid closing lots to then reopen lots.

Change-Id: I26ca12b60b98c1574ff15a86fa1f124edd169ad2
Reviewed-on: http://review.couchbase.org/90145
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>