MB-27415 implement an asynchronous overflow connection closer
As per tests with Pavel, it really is the overflow connections that are constantly deplenting local ports.
Introduced an asynchronous closer, which can be turned on on demand, which monitors the
connections channel once every 30 seconds, and if connections have been requested at
a rate of less than once per 1ms, closes the existing overflow connections.
This behavior is turned on specifically for N1QL and nothing else.
Also added a mechanism to turn off overflow connections for good (need a new function
because traditionally SetConnectionPoolParams(n, 0) means don't change overflow connections)

Change-Id: I1da6a5181db2b0b8379c87429a4e78002536a02e
Reviewed-on: http://review.couchbase.org/90346
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>