MB-32301. During KV fetch ON connection error retry non-bulkGet
Fetch more than one key uses doBulkGet()
All others uses Do() (Fetch one key or INSERT/UPDATE/DELETE,....)

The connection might have disconnected by memached.
This can happen while we are using or connection is in the idle in the pool
The connection needs to handle any of this transient errors and retry. We already
do this in doBulkGet(), Not in Do().

The DoNoDeadline() code is same as Do() except caller takes care of deadline setting.
Combined both by passing the flag and removed DoNoDeadline()

Change-Id: If8810245556d218a4daac713b435fcaab3b4eb22
Reviewed-on: http://review.couchbase.org/102648
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>