MB-33208 delay connPool closing on pool closing
With the fix to M B 33185, we close the connections when we close the pool,
however there may still be active bulk gets.
Use the Bucket finalizer to close the connection pools, and use the pool Close method
to signal that the finalizer is not operating on a bucket still in use.

Change-Id: Id0bec85376bbdd2001f158952fe5a256a00f2fe6
Reviewed-on: http://review.couchbase.org/105596
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>