MB-37493 improve deadline load
Currently in go-couchbase we set and reset deadlines for every operation.
net package deadline handling is not great (10% cpu load), so we seek to cut down.

1st change is to make sure that every request has a deadline set (even if 0), rather than having each
request with a deadline do it twice (note that the default deadline is 120 secs, so this means that all
requests have a deadline anyway

2nd change is to be slightly more sloppy with deadlines: rather than setting deadlines with a nanosecond precision,
we get the current time to the second.
This means that for a whole second, a connection has uses the same deadline, which means we don't have to set it again
(this works in conjunction with the gomemcached change).

The two changes together mean that for Q1 we set Get() deadlines 1000 times per second, as opposed to
280000 times.

Latency improves to 0.8ms (first time we are sub millisecond at full speed), ad throughput to 148k

Change-Id: I703956278c536cb0964c20086811dae9544ab6d1
Reviewed-on: http://review.couchbase.org/124172
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>