MB-41540 scope with % is mishandled
The bug is a bit misleading, in that any kv object with a % is mishandled: buckets, scopes and collections.
This change fixes the URI issues in object names.
Note that bucket names with % were never ever tested, and I have just seen a few panics
due to such buckets not being loaded properly.
Expect more issues...

Change-Id: I72dde78aa71fa872bc17613bbbbae5283308f8f4
Reviewed-on: http://review.couchbase.org/c/go-couchbase/+/136882
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>